[nanvix-terminal] E: Split by deploy
Split the terminal into per-deployment-mode modules.

The monolithic lib.rs implementation is replaced by three
deployment-specific modules. In standalone mode, the terminal
drives a User VM instance directly via StandaloneVmHandle,
bypassing the sandbox cache, gateway sockets, and control-plane.
In single-process mode, it uses the SimpleSandboxCache. In
multi-process mode, it connects through the full SandboxCache.

Changes:
- Cargo.toml: make nanvix-sandbox, nanvix-sandbox-cache, config,
  and uservm optional. Route features to appropriate deps.
  Use ? syntax for optional feature forwarding.
- lib.rs: replace monolith with conditional module routing
  and re-exports. Add compile_error for conflicting features.
- multi_process.rs: terminal via sandbox cache gateway sockets.
- single_process.rs: terminal via embedded linuxd.
- standalone.rs: terminal via StandaloneVmHandle (direct VM).
